+++ R/JackSymPol.R
@@ -1,0 +1,39 @@
+#' Jack polynomial with symbolic Jack parameter
+#'
+#' Returns the Jack polynomial with symbolic Jack parameter.
+#'
+#' @param n number of variables, a positive integer
+#' @param lambda an integer partition, given as a vector of decreasing
+#'   integers
+#' @param which which Jack polynomial, \code{"J"}, \code{"P"}, \code{"Q"},
+#'  or \code{"C"}
+#'
+#' @return A \code{symbolicQspray} object.
+#'
+#' @export
+#' @importFrom symbolicQspray symbolicQspray_from_list
+#'
+#' @examples
+#' JackSymPol(3, lambda = c(3, 1))
+JackSymPol <- function(n, lambda, which = "J") {
+  stopifnot(isPositiveInteger(n), isPartition(lambda))
+  lambda <- as.integer(lambda[lambda != 0])
+  which <- match.arg(which, c("J", "P", "Q", "C"))
+  JackPolynomial <- symbolicQspray_from_list(
+    JackSymPolRcpp(as.integer(n), lambda)
+  )
+  if(which != "J") {
+    if(which == "C") {
+      K <- symbolicJackCcoefficient(lambda)
+      JackPolynomial <- K * JackPolynomial
+    } else {
+      invK <- switch(
+        which,
+        "P" = symbolicJackPcoefficientInverse(lambda),
+        "Q" = symbolicJackQcoefficientInverse(lambda)
+      )
+      JackPolynomial <- JackPolynomial / invK
+    }
+  }
+  JackPolynomial
+}