+++ NEWS.md
@@ -1,3 +1,35 @@
+# Version 6.0.0 (2024-05-XX)
+
+- It is now possible to get a Jack polynomial with a symbolic Jack parameter 
+in its coefficients. Such polynomials are returned by the function `JackSymPol`. 
+This big progress is the reason for which I increased the major component of 
+the version of the package.
+
+- Since the functions implemented with **Rcpp** are highly more efficient, 
+the functions `Jack`, `JackPol`, `Schur`, `SchurPol`, `Zonal`, `ZonalPol`, 
+`ZonalQ`, `ZonalQPol`, have been renamed to `JackR`, `JackPolR`, `SchurR`,
+`SchurPolR`, `ZonalR`, `ZonalPolR`, `ZonalQR`, `ZonalQPolR`, and the functions 
+`JackCPP`, `JackPolCPP`, `SchurCPP`, `SchurPolCPP`, `ZonalCPP`, `ZonalPolCPP`, 
+`ZonalQCPP`, `ZonalQPolCPP` have been renamed to `Jack`, `JackPol`, `Schur`,
+`SchurPol`, `Zonal`, `ZonalPol`, `ZonalQ`, `ZonalQPol`.
+
+- New function `LRmult`, to compute the expression of the product of two Schur 
+polynomials as a linear combination of Schur polynomials, using the 
+Littlewood-Richardson rule.
+
+- New function `LRskew`, to compute the expression of a skew Schur 
+polynomial as a linear combination of Schur polynomials, using the 
+Littlewood-Richardson rule.
+
+- Based on `LRskew`, the new function `SkewSchurPol` computes the skew Schur 
+polynomial of a given skew partition.
+
+- Actually there are four possible Jack polynomials of a given partition for a
+given `alpha`, denoted by `J`, `C`, `Q` or `P` (they are the same up to a 
+normalization constant). It is now possible to get any of them (the previous
+versions only allowed to get the `J` polynomial).
+
+
 # Version 5.3.0 (2023-07-04)
 
 The Julia stuff has been removed.